+++ src/repr/col_sheet.rs
@@ -1,19 +1,22 @@
-#![allow(unused_imports, dead_code)]
 use csv::{ReaderBuilder, Trim};
-use std::{
-    collections::{HashMap, HashSet},
-    fmt::Debug,
-    path::{Path, PathBuf},
-    slice::{Iter, IterMut},
-    str::FromStr,
-};
+use std::{fmt::Debug, path::Path, slice::Iter, str::FromStr};
 
+#[allow(unused_imports)]
 use crate::models::{
     bar::{Bar, BarChart},
     line::{Line, LineGraph},
     stacked_bar::{StackedBar, StackedBarChart},
     Point, Scale, ScaleKind,
 };
+
+mod arraytext;
+pub use arraytext::*;
+
+mod arrayi32;
+pub use arrayi32::*;
+
+mod arrayisize;
+pub use arrayisize::*;
 
 use super::builders::SheetBuilder;
 use super::utils::{HeaderLabelStrategy, HeaderTypesStrategy};
@@ -113,375 +116,8 @@ pub trait Column: Sealed + Debug {
 
 #[derive(Debug)]
 pub struct ColumnHeader<'a> {
-    header: Option<&'a String>,
-    kind: DataType,
-}
-
-#[derive(Debug, Clone, PartialEq, Default)]
-pub struct ArrayI32 {
-    header: Option<String>,
-    cells: Vec<Option<i32>>,
-}
-
-impl ArrayI32 {
-    fn new() -> Self {
-        Self::default()
-    }
-
-    fn from_iterator(values: impl Iterator<Item = i32>) -> Self {
-        Self {
-            cells: values.map(Some).collect(),
-            ..Default::default()
-        }
-    }
-
-    fn from_iterator_option(values: impl Iterator<Item = Option<i32>>) -> Self {
-        Self {
-            cells: values.collect(),
-            ..Default::default()
-        }
-    }
-
-    fn set_header(&mut self, header: String) -> &mut Self {
-        self.header = Some(header);
-        self
-    }
-
-    fn get(&self, idx: usize) -> Option<i32> {
-        self.cells.get(idx)?.as_ref().copied()
-    }
-
-    fn iter(&self) -> Iter<'_, Option<i32>> {
-        self.cells.iter()
-    }
-
-    fn parse_str(values: &Vec<String>) -> Option<Self> {
-        let mut cells = Vec::default();
-
-        for value in values {
-            let value = parse_helper::<i32>(value).ok()?;
-            cells.push(value)
-        }
-
-        Some(Self {
-            header: None,
-            cells,
-        })
-    }
-}
-
-impl Sealed for ArrayI32 {
-    fn push(&mut self, value: &str) {
-        let value = match parse_helper::<i32>(value) {
-            Ok(val) => val,
-            Err(_) => None,
-        };
-        self.cells.push(value)
-    }
-
-    fn remove(&mut self, idx: usize) {
-        if idx >= self.len() {
-            return;
-        }
-        self.cells.remove(idx);
-    }
-
-    fn insert(&mut self, value: &str, idx: usize) {
-        if idx >= self.len() {
-            return;
-        }
-
-        let Ok(parsed) = parse_helper::<i32>(value) else {
-            return;
-        };
-
-        self.cells.insert(idx, parsed);
-    }
-}
-
-impl Column for ArrayI32 {
-    fn label(&self) -> Option<&String> {
-        self.header.as_ref()
-    }
-
-    fn kind(&self) -> DataType {
-        DataType::Int32
-    }
-
-    fn len(&self) -> usize {
-        self.cells.len()
-    }
-
-    fn set_header(&mut self, header: String) {
-        self.header = Some(header);
-    }
-
-    fn set_position(&mut self, value: &str, idx: usize) {
-        let Ok(parsed) = parse_helper::<i32>(value) else {
-            return;
-        };
-
-        let Some(prev) = self.cells.get_mut(idx) else {
-            return;
-        };
-
-        *prev = parsed;
-    }
-
-    fn swap(&mut self, x: usize, y: usize) {
-        if x >= self.len() || y >= self.len() {
-            return;
-        }
-
-        self.cells.swap(x, y);
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Default)]
-pub struct ArrayISize {
-    header: Option<String>,
-    cells: Vec<Option<isize>>,
-}
-
-impl ArrayISize {
-    fn new() -> Self {
-        Self::default()
-    }
-
-    fn from_iterator(values: impl Iterator<Item = isize>) -> Self {
-        Self {
-            cells: values.map(Some).collect(),
-            ..Default::default()
-        }
-    }
-
-    fn from_iterator_option(values: impl Iterator<Item = Option<isize>>) -> Self {
-        Self {
-            cells: values.collect(),
-            ..Default::default()
-        }
-    }
-
-    fn set_header(&mut self, header: String) -> &mut Self {
-        self.header = Some(header);
-        self
-    }
-
-    fn get(&self, idx: usize) -> Option<isize> {
-        self.cells.get(idx)?.as_ref().copied()
-    }
-
-    fn iter(&self) -> Iter<'_, Option<isize>> {
-        self.cells.iter()
-    }
-
-    fn parse_str(values: &Vec<String>) -> Option<Self> {
-        let mut cells = Vec::default();
-
-        for value in values {
-            let value = parse_helper::<isize>(value).ok()?;
-            cells.push(value)
-        }
-
-        Some(Self {
-            header: None,
-            cells,
-        })
-    }
-}
-
-impl Sealed for ArrayISize {
-    fn push(&mut self, value: &str) {
-        let value = match parse_helper::<isize>(value) {
-            Ok(val) => val,
-            Err(_) => None,
-        };
-        self.cells.push(value)
-    }
-
-    fn remove(&mut self, idx: usize) {
-        if idx >= self.len() {
-            return;
-        }
-        self.cells.remove(idx);
-    }
-
-    fn insert(&mut self, value: &str, idx: usize) {
-        if idx >= self.len() {
-            return;
-        }
-
-        let Ok(parsed) = parse_helper::<isize>(value) else {
-            return;
-        };
-
-        self.cells.insert(idx, parsed);
-    }
-}
-
-impl Column for ArrayISize {
-    fn len(&self) -> usize {
-        self.cells.len()
-    }
-
-    fn kind(&self) -> DataType {
-        DataType::ISize
-    }
-
-    fn label(&self) -> Option<&String> {
-        self.header.as_ref()
-    }
-
-    fn set_header(&mut self, header: String) {
-        self.header = Some(header)
-    }
-
-    fn set_position(&mut self, value: &str, idx: usize) {
-        let Ok(parsed) = parse_helper::<isize>(value) else {
-            return;
-        };
-
-        let Some(prev) = self.cells.get_mut(idx) else {
-            return;
-        };
-
-        *prev = parsed;
-    }
-
-    fn swap(&mut self, x: usize, y: usize) {
-        if x >= self.len() || y >= self.len() {
-            return;
-        }
-
-        self.cells.swap(x, y)
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Default)]
-pub struct ArrayText {
-    header: Option<String>,
-    cells: Vec<Option<String>>,
-}
-
-impl ArrayText {
-    fn new() -> Self {
-        Self::default()
-    }
-
-    fn from_iterator(values: impl Iterator<Item = String>) -> Self {
-        Self {
-            cells: values.map(Some).collect(),
-            ..Default::default()
-        }
-    }
-
-    fn from_iterator_option(values: impl Iterator<Item = Option<String>>) -> Self {
-        Self {
-            cells: values.collect(),
-            ..Default::default()
-        }
-    }
-
-    fn set_header(&mut self, header: impl Into<String>) -> &mut Self {
-        self.header = Some(header.into());
-        self
-    }
-
-    fn get(&self, idx: usize) -> Option<String> {
-        self.cells.get(idx)?.as_ref().cloned()
-    }
-
-    fn get_ref(&self, idx: usize) -> Option<&String> {
-        self.cells.get(idx)?.as_ref()
-    }
-
-    fn iter(&self) -> Iter<'_, Option<String>> {
-        self.cells.iter()
-    }
-
-    fn parse_str(values: Vec<String>) -> Self {
-        let mut cells = Vec::default();
-
-        for value in values {
-            // Always successful
-            let value = match parse_helper::<String>(&value) {
-                Ok(val) => val,
-                Err(_) => None,
-            };
-            cells.push(value);
-        }
-
-        Self {
-            header: None,
-            cells,
-        }
-    }
-}
-
-impl Sealed for ArrayText {
-    fn push(&mut self, value: &str) {
-        let value = match parse_helper::<String>(value) {
-            Ok(val) => val,
-            Err(_) => None,
-        };
-        self.cells.push(value)
-    }
-
-    fn remove(&mut self, idx: usize) {
-        if idx >= self.len() {
-            return;
-        }
-        self.cells.remove(idx);
-    }
-
-    fn insert(&mut self, value: &str, idx: usize) {
-        if idx >= self.len() {
-            return;
-        }
-        let Ok(parsed) = parse_helper::<String>(value) else {
-            return;
-        };
-
-        self.cells.insert(idx, parsed);
-    }
-}
-
-impl Column for ArrayText {
-    fn len(&self) -> usize {
-        self.cells.len()
-    }
-
-    fn kind(&self) -> DataType {
-        DataType::Text
-    }
-
-    fn label(&self) -> Option<&String> {
-        self.header.as_ref()
-    }
-
-    fn set_header(&mut self, header: String) {
-        self.set_header(header);
-    }
-
-    fn set_position(&mut self, value: &str, idx: usize) {
-        let Ok(parsed) = parse_helper::<String>(value) else {
-            return;
-        };
-
-        let Some(prev) = self.cells.get_mut(idx) else {
-            return;
-        };
-
-        *prev = parsed;
-    }
-
-    fn swap(&mut self, x: usize, y: usize) {
-        if x >= self.len() || y >= self.len() {
-            return;
-        }
-
-        self.cells.swap(x, y)
-    }
+    pub header: Option<&'a String>,
+    pub kind: DataType,
 }
 
 pub struct ColumnSheet {
