+++ src/repr/col_sheet/arrayi32.rs
@@ -1,0 +1,121 @@
+use super::{parse_helper, Column, DataType, Iter, Sealed};
+
+#[derive(Debug, Clone, PartialEq, Default)]
+pub struct ArrayI32 {
+    header: Option<String>,
+    cells: Vec<Option<i32>>,
+}
+
+impl ArrayI32 {
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    pub fn from_iterator(values: impl Iterator<Item = i32>) -> Self {
+        Self {
+            cells: values.map(Some).collect(),
+            ..Default::default()
+        }
+    }
+
+    pub fn from_iterator_option(values: impl Iterator<Item = Option<i32>>) -> Self {
+        Self {
+            cells: values.collect(),
+            ..Default::default()
+        }
+    }
+
+    pub fn set_header(&mut self, header: String) -> &mut Self {
+        self.header = Some(header);
+        self
+    }
+
+    pub fn get(&self, idx: usize) -> Option<i32> {
+        self.cells.get(idx)?.as_ref().copied()
+    }
+
+    pub fn iter(&self) -> Iter<'_, Option<i32>> {
+        self.cells.iter()
+    }
+
+    pub fn parse_str(values: &Vec<String>) -> Option<Self> {
+        let mut cells = Vec::default();
+
+        for value in values {
+            let value = parse_helper::<i32>(value).ok()?;
+            cells.push(value)
+        }
+
+        Some(Self {
+            header: None,
+            cells,
+        })
+    }
+}
+
+impl Sealed for ArrayI32 {
+    fn push(&mut self, value: &str) {
+        let value = match parse_helper::<i32>(value) {
+            Ok(val) => val,
+            Err(_) => None,
+        };
+        self.cells.push(value)
+    }
+
+    fn remove(&mut self, idx: usize) {
+        if idx >= self.len() {
+            return;
+        }
+        self.cells.remove(idx);
+    }
+
+    fn insert(&mut self, value: &str, idx: usize) {
+        if idx >= self.len() {
+            return;
+        }
+
+        let Ok(parsed) = parse_helper::<i32>(value) else {
+            return;
+        };
+
+        self.cells.insert(idx, parsed);
+    }
+}
+
+impl Column for ArrayI32 {
+    fn label(&self) -> Option<&String> {
+        self.header.as_ref()
+    }
+
+    fn kind(&self) -> DataType {
+        DataType::Int32
+    }
+
+    fn len(&self) -> usize {
+        self.cells.len()
+    }
+
+    fn set_header(&mut self, header: String) {
+        self.header = Some(header);
+    }
+
+    fn set_position(&mut self, value: &str, idx: usize) {
+        let Ok(parsed) = parse_helper::<i32>(value) else {
+            return;
+        };
+
+        let Some(prev) = self.cells.get_mut(idx) else {
+            return;
+        };
+
+        *prev = parsed;
+    }
+
+    fn swap(&mut self, x: usize, y: usize) {
+        if x >= self.len() || y >= self.len() {
+            return;
+        }
+
+        self.cells.swap(x, y);
+    }
+}
